+++ src/viriatum/test/simple_test.c
@@ -1490,6 +1490,8 @@ static struct test_entry_t _simple_entries[] = {
     V_TEST_T(test_dev_options_service, "service"),
     V_TEST_T(test_print_config_service, "service"),
     V_TEST_T(test_list_handlers_service, "service"),
+    V_TEST_T(test_check_service, "service"),
+    V_TEST_T(test_handlers_service, "service"),
     V_TEST_T(test_ephemeral_service, "service"),
     V_TEST_T(test_polling_service, "service"),
     V_TEST_T(test_polling_connection, "polling"),